+++ lib/open_api_typesense/schemas/collection_response.ex
@@ -13,6 +13,7 @@ defmodule OpenApiTypesense.CollectionResponse do
           name: String.t(),
           num_documents: integer,
           symbols_to_index: [String.t()],
+          synonym_sets: [String.t()],
           token_separators: [String.t()],
           voice_query_model: OpenApiTypesense.VoiceQueryModelCollectionConfig.t()
         }
@@ -23,6 +24,7 @@ defmodule OpenApiTypesense.CollectionResponse do
     :metadata,
     :name,
     :num_documents,
+    :synonym_sets,
     default_sorting_field: "",
     enable_nested_fields: false,
     symbols_to_index: [],
@@ -80,6 +82,7 @@ defmodule OpenApiTypesense.CollectionResponse do
       name: {:string, :generic},
       num_documents: :integer,
       symbols_to_index: [string: :generic],
+      synonym_sets: [string: :generic],
       token_separators: [string: :generic],
       voice_query_model: {OpenApiTypesense.VoiceQueryModelCollectionConfig, :t}
     ]