+++ test/operations/override_test.exs
@@ -14,7 +14,7 @@ defmodule OverrideTest do
 
   @tag ["28.0": true, "27.1": true, "27.0": true, "26.0": true]
   test "error: retrieve an override", %{conn: conn, map_conn: map_conn} do
-    assert {:error, %ApiResponse{message: "Not Found"}} =
+    assert {:error, %ApiResponse{message: "Collection not found"}} =
              Override.get_search_override("helmets", "custom-helmet")
 
     assert {:error, _} = Override.get_search_override("helmets", "custom-helmet", [])